First step into translatable models
This management command dumps all values that should be translated from
the database models into a python file. The strings are marked with
ugettext and the regular compilemessages from django-admin creates the
pofile for translation. This is located on a separate app. Next step is
to sync the pofile with the actual translation fields in the database.